+++ internal/resolver/storage.go
@@ -12,16 +12,19 @@ import (
 
 // CachingStorage implements resolver for storages, caching the results
 type CachingStorage struct {
-	cachedStorages *upcloud.Storages
+	Cache[upcloud.Storage]
 }
 
-// make sure we implement the ResolutionProvider interface
-var _ ResolutionProvider = &CachingStorage{}
+// make sure we implement the ResolutionProvider interfaces
+var (
+	_ ResolutionProvider                         = &CachingStorage{}
+	_ CachingResolutionProvider[upcloud.Storage] = &CachingStorage{}
+)
 
-func storageMatcher(cached []upcloud.Storage) func(arg string) Resolved {
+func (s *CachingStorage) matchCached() func(arg string) Resolved {
 	return func(arg string) Resolved {
 		rv := Resolved{Arg: arg}
-		for _, storage := range cached {
+		for _, storage := range s.cache {
 			rv.AddMatch(storage.UUID, MatchTitle(arg, storage.Title))
 			rv.AddMatch(storage.UUID, MatchUUID(arg, storage.UUID))
 		}
@@ -31,12 +34,16 @@ func storageMatcher(cached []upcloud.Storage) func(arg string) Resolved {
 
 // Get implements ResolutionProvider.Get
 func (s *CachingStorage) Get(ctx context.Context, svc internal.AllServices) (Resolver, error) {
-	var err error
-	s.cachedStorages, err = svc.GetStorages(ctx, &request.GetStoragesRequest{})
+	storages, err := svc.GetStorages(ctx, &request.GetStoragesRequest{})
 	if err != nil {
 		return nil, err
 	}
-	return storageMatcher(s.cachedStorages.Storages), nil
+
+	for _, storage := range storages.Storages {
+		s.AddCached(storage.UUID, storage)
+	}
+
+	return s.matchCached(), nil
 }
 
 // PositionalArgumentHelp implements resolver.ResolutionProvider
@@ -46,23 +53,10 @@ func (s *CachingStorage) PositionalArgumentHelp() string {
 
 // Resolve is a helper method for commands to resolve networks inside Execute(), outside arguments
 func (s *CachingStorage) Resolve(arg string) (resolved string, err error) {
-	if s.cachedStorages == nil {
+	if s.cache == nil {
 		return "", errors.New("caching storage does not have a cache initialized")
 	}
 
-	r := storageMatcher(s.cachedStorages.Storages)(arg)
+	r := s.matchCached()(arg)
 	return r.GetOnly()
-}
-
-// GetCached is a helper method for commands to use when they need to get an item from the cached results
-func (s *CachingStorage) GetCached(uuid string) (upcloud.Storage, error) {
-	if s.cachedStorages == nil {
-		return upcloud.Storage{}, errors.New("caching storage does not have a cache initialized")
-	}
-	for _, storage := range s.cachedStorages.Storages {
-		if storage.UUID == uuid {
-			return storage, nil
-		}
-	}
-	return upcloud.Storage{}, NotFoundError(uuid)
 }